requirements.txt
# note which revision of python, for example 3.9.6
# in this file, insert all the pip install needs, include revision
# python 3.11

langchain==0.3.18
langchain-openai==0.3.5
langchain-anthropic==0.3.7
langgraph==0.2.72
langmem==0.0.8
python-dotenv==1.0.1
sentence-transformers==2.2.2
langchain-google-genai==0.0.2